Was added log4j support. Some refactoring

+++ src/main/java/payments/dao/BankAccountDao.java
@@ -5,6 +5,6 @@ import payments.model.entity.BankAccount;
 import java.util.Optional;
 
 public interface BankAccountDao extends CommonDao<BankAccount> {
-    Optional<BankAccount> findBankAccountByCard(int id);
+    Optional<BankAccount> findBankAccountByCard(long id);
     Optional<BankAccount> findBankAccountByNumber(String number);
 }
